fix: upgrade telemetry packetId to BIGINT for PostgreSQL/MySQL
Meshtastic packet IDs are unsigned 32-bit values (up to ~4.3 billion)
but PostgreSQL/MySQL INTEGER is signed 32-bit (max ~2.1 billion),
causing insert failures for large packetId values. Adds migration 075
to ALTER the column type and updates schema/CREATE scripts.

+++ src/db/schema/mysql-create.ts
@@ -115,7 +115,7 @@ export const MYSQL_SCHEMA_SQL = `
     unit VARCHAR(255),
     createdAt BIGINT NOT NULL,
     packetTimestamp BIGINT,
-    packetId INT,
+    packetId BIGINT,
     channel INT,
     precisionBits INT,
     gpsAccuracy DOUBLE,
